64-bit compatibility fixes in userland/tests

+++ tests/block_test.c
@@ -25,7 +25,7 @@ void *block_thread(void* arg)
 		printf_safe("[A] pthread %d on vcore %d\n", pthread_self()->id, vcore_id());
 		sys_block(5000 + pthread_self()->id);
 	}
-	return (void*)(pthread_self()->id);
+	return (void*)(long)pthread_self()->id;
 }
 
 int main(int argc, char** argv)
